requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.0
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
audioop-lts==0.2.1
black==25.1.0
Brotli==1.1.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cohere==5.15.0
cssselect2==0.8.0
discord.py==2.5.2
fastapi==0.115.12
fastavro==1.11.1
filelock==3.18.0
fonttools==4.58.0
frozenlist==1.6.0
fsspec==2025.5.1
h11==0.16.0
hf-xet==1.1.2
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.32.0
idna==3.10
markdown2==2.5.3
multidict==6.4.4
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.3.8
propcache==0.3.1
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
pydyf==0.11.0
PyPDF2==3.0.1
pyphen==0.17.2
python-dotenv==1.1.0
PyYAML==6.0.2
requests==2.32.3
sniffio==1.3.1
starlette==0.46.2
tinycss2==1.4.0
tinyhtml5==2.0.0
tokenizers==0.21.1
tqdm==4.67.1
types-requests==2.32.0.20250515
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
weasyprint==65.1
webencodings==0.5.1
yarl==1.20.0
zopfli==0.2.3.post1
